feat(recipe): 处方生成查看

+++ database/migrations/2021_02_05_092056_create_foods_table.php
@@ -18,20 +18,20 @@ class CreateFoodsTable extends Migration
             // 食品名称
             $table->string('name');
             // 食品GI
-            $table->integer('gi');
+            $table->float('gi');
             // 食品类型
             // 1 表示主食，2 表示奶类，3 表示蔬菜菌类，4 表示豆制品，
             // 5 表示豆制品，6 表示饮料，7 表示油脂类，8 表示调味品，
             // 9 表示零食点心，10 表示其他，11 表示菜肴
             // $table->enum('type',[1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11]);
             // 每一百克卡路里
-            $table->integer('energy');
+            $table->float('energy');
             // 碳水化合物（g）
-            $table->integer('carbohydrate');
+            $table->float('carbohydrate');
             // 脂肪（g）
-            $table->integer('axunge');
+            $table->float('axunge');
             // 蛋白质（g）
-            $table->integer('protein');
+            $table->float('protein');
             $table->timestamps();
         });
     }
